refactor: reorganize create-table code and move tests to integration tests
Address review comments:
- Move CreateTableTransactionBuilder and all validation helpers to
  transaction/builder/create_table.rs to separate builder logic from
  the type definition and entry point.
- Move try_new_create_table constructor from mod.rs to
  transaction/create_table.rs where it belongs conceptually.
- Move CreateTableTransaction type alias and docs from mod.rs to
  transaction/create_table.rs (re-exported so import paths still work).
- Move test_create_table_txn_debug test from unit tests (mod.rs) to
  integration tests (kernel/tests/create_table.rs).
- Remove internal test_table_setup helper from kernel/src/utils.rs
  (no longer needed; integration tests use test_utils crate instead).
- Restore test_validate_blind_append_rejects_create_table in mod.rs
  to preserve coverage of runtime blind-append rejection.
- No functional changes.
