+++ kernel/src/transaction/builder/mod.rs
@@ -1,0 +1,8 @@
+//! Builder modules for transaction construction.
+
+// Allow `pub` items in this module even though the module itself may be `pub(crate)`.
+// The module visibility controls external access; items are `pub` for use within the crate
+// and for tests. Also allow dead_code since these are used by integration tests.
+#![allow(unreachable_pub, dead_code)]
+
+pub mod create_table;
